refactor(log): scaffold modules/log and migrate log components with proper re-exports
