+++ src/components/sections/log/index.ts
@@ -1,0 +1,8 @@
+export * from "@/modules/log/components/LogIndex";
+export * from "@/modules/log/components/mocks";
+export { default as ArticleHeader } from "@/modules/log/components/ArticleDetail/ArticleHeader";
+export { default as ArticleBody } from "@/modules/log/components/ArticleDetail/ArticleBody";
+export { default as AuthorCard } from "@/modules/log/components/ArticleDetail/AuthorCard";
+export { default as RelatedArticles } from "@/modules/log/components/ArticleDetail/RelatedArticles";
+export { default as Comments } from "@/modules/log/components/ArticleDetail/Comments";
+export * from "@/modules/log/components/ArticleDetail/mocks";
