+++ src/modules/log/components/ArticleDetail/ArticleHeader.tsx
@@ -66,3 +66,5 @@ export default function ArticleHeader({
     </div>
   )
 }
+
+
