+++ src/modules/log/components/ArticleDetail/AuthorCard.tsx
@@ -51,3 +51,5 @@ export default function AuthorCard({
     </div>
   )
 }
+
+
