+++ src/modules/log/components/ArticleDetail/Comments.tsx
@@ -61,3 +61,5 @@ export default function Comments({ comments, onSubmit }: CommentsProps) {
     </div>
   )
 }
+
+
