+++ src/modules/log/components/ArticleDetail/RelatedArticles.tsx
@@ -40,3 +40,5 @@ export default function RelatedArticles({ articles }: RelatedArticlesProps) {
     </div>
   )
 }
+
+
