+++ src/modules/log/components/ArticleDetail/mocks.tsx
@@ -54,4 +54,5 @@ export const mockArticle = {
       console.log("User submitted:", comment)
     },
   }
-  
+
+
