+++ src/modules/log/components/LogIndex.tsx
@@ -58,3 +58,4 @@ export function LogIndex() {
     </>
   );
 }
+
