+++ src/modules/log/components/mocks.tsx
@@ -83,3 +83,5 @@ export const mockArticles: LogArticleCard[] = [
     tags: ["Testing", "JavaScript"],
   },
 ];
+
+
